// src/shared/api/generated/model/accountItemResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { AccountItemResponseAccountType } from './accountItemResponseAccountType';
import type { AccountItemResponseStatus } from './accountItemResponseStatus';

export interface AccountItemResponse {
  /** 계좌 내부 식별자 */
  accountId?: number;
  /** 화면에 표시할 계좌명 */
  accountName?: string;
  /** 계좌번호 */
  accountNumber?: string;
  /** 계좌 유형 */
  accountType?: AccountItemResponseAccountType;
  /** 계좌 잔액(원) */
  balance?: number;
  /** 계좌 상태 */
  status?: AccountItemResponseStatus;
  /** 계좌 개설일 */
  openedDate?: string;
  /**
     * 마지막 거래 일시. 거래내역이 없으면 null
     * @nullable
     */
  lastTransactionAt?: string | null;
  /**
     * 만기일. 입출금계좌는 null
     * @nullable
     */
  maturityDate?: string | null;
  /** 이체 가능 여부 */
  transferEnabled?: boolean;
}